datapath: Support NodePort BPF on L2-less devices
This commit extends NodePort BPF by making it possible to run it on
L3 network devices (without L2 addr). One prominent case is the
Wireguard tunnel device (wg0).

The main idea of the change is to make ETH_HLEN configurable via ELF
templating (on L2-less devices we set it to 0 during the load time),
and to craft an L2 hdr when forwarding from L2-less to L2 device.

So far I've manually tested the following NodePort BPF (SNAT) scenarios:

- NP request is sent from outside to eth0 (wg0 is direct-routing dev).
- NP request is sent from outside to wg0 (wg0 is direct-routing dev).

Signed-off-by: Martynas Pumputis <m@lambda.lt>