Update PR triggers and k8s/kernel pairs
Following today's cartography and multiple fixes, documentation has been
updated accordingly.

Context

When splitting `ginkgo-kubernetes-all.Jenkinsfile` in #14861, we
offloaded the actual work to PR jobs. However, this means daily
`cilium-master-K8s-all`-triggered job results were mixed in with
PR-triggered job results. For tracking purposes, separate jobs needed to
be setup.

We took this oppurtunity to review, document, and accordingly fix a big
part of the CI jobs, notably:
- What to test automatically at regular intervals.
- Which jobs are required for which PRs.
- Which phrases trigger which jobs.

Detailed recap of changes made so far:
- Creation and publication of The Table of Truth™️
  https://docs.google.com/spreadsheets/d/1TThkqvVZxaqLR-Ela4ZrcJ0lrTJByCqrbdCjnI32_X0
- Master jobs
  - Automated testing jobs running every 4 hours or daily have been
    setup according to the Table of Truth™️.
  - Most jobs have been renamed to make it clear which K8s version and
    Kernel version they're running.
  - `cilium-master-K8s-all`, now superseded by individual jobs, has been
    disabled and moved to the `Obsolete` view.
- PR jobs
  - All jobs marked as required for a PR in the Table of Truth™️ have
    been setup accordingly on GitHub.
  - All the triggers in required PR jobs have been reviewed, fixed
    according to the Table of Truth™️, and documented below the Table of
    Truth™️.
  - `Cilium-PR-Ginkgo-Tests-Kernel` (job for K8s 1.19 / Kernel 4.19)
    edited to correctly use K8s version 1.19 (and not 1.18, which seems
    to be a leftover).

Note: unlike 1.8 and 1.9, `test-backport-1.7` does not trigger
individual jobs. It still triggers the old `Cilium-PR-Ginkgo-Tests-K8s`,
because we were unsure of the robustness of `ginkgo-kernel.Jenkinsfile`
on 1.7. We may backport the updated `ginkgo-kernel.Jenkinsfile` to 1.7
in a future PR.

Signed-off-by: Nicolas Busseneau <nicolas@isovalent.com>